fix: code review fixes for story 7-3-integration-tests-adr003
- Move _seed_integration_catalog from conftest.py to _helpers.py to avoid
  direct conftest imports (anti-pattern)
- Add try/finally blocks for actor system shutdown in inline fixture tests
  to prevent resource leaks on assertion failure
- Strengthen event assertion in test_send_message_routes_through_team_handle
  to verify sent message content appears in events

Relates to #75
